Fourth commit

+++ src/test/java/testBase/BaseClass.java
@@ -6,6 +6,9 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.List;
 import java.io.IOException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
 import java.time.Duration;
 import java.util.Properties;
 
@@ -13,24 +16,27 @@ import org.apache.commons.text.RandomStringGenerator;
 import org.apache.logging.log4j.Logger;
 import org.apache.logging.log4j.LogManager;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.Platform;
 import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 import org.openqa.selenium.edge.EdgeDriver;
 import org.openqa.selenium.firefox.FirefoxDriver;
+import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.testng.annotations.AfterClass;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Parameters;
 
 public class BaseClass {
 
-	public WebDriver driver;
+	public static WebDriver driver;
 	public Logger logger;
 	public Properties p;
 
 	@BeforeClass(groups= {"Sanity", "Regression", "Datadriven", "Master"})
 	@Parameters({"os","browser"})
-	public void setup(String os, String browser) throws IOException {
+	public void setup(String os, String browser) throws IOException, URISyntaxException {
 		
 		//Loading config.properties file
 		FileReader file = new FileReader(".//src//test//resources//config.properties");
@@ -39,12 +45,46 @@ public class BaseClass {
 		
 		logger = LogManager.getLogger(this.getClass());
 		
-		switch(browser.toLowerCase()) {
-		case "chrome" : driver = new ChromeDriver(); break;
-		case "edge" : driver = new EdgeDriver(); break;
-		case "firefox" : driver = new FirefoxDriver(); break;
-		default : System.out.println("Invalid browser"); return;
+		if(p.getProperty("execution_env").equalsIgnoreCase("remote")) {
+			DesiredCapabilities capabilities = new DesiredCapabilities();
+			
+			//OS
+			if(os.equalsIgnoreCase("windows")) {
+				capabilities.setPlatform(Platform.WIN11);
+			}
+			else if(os.equalsIgnoreCase("mac")) {
+				capabilities.setPlatform(Platform.MAC);
+			}
+			else {
+				System.out.println("No matching os");
+				return;
+			}
+			
+			//Browser
+			switch(browser.toLowerCase()) {
+			case "chrome": capabilities.setBrowserName("chrome"); break;
+			case "edge": capabilities.setBrowserName("MicrosoftEdge"); break;
+			default: System.out.println("No matching browser"); return;
+			
+			
+			}
+			
+			driver = new RemoteWebDriver(new URI("http://192.168.4.221:4444/wd/hub").toURL(), capabilities);
+	
+			
+			
 		}
+		
+		if(p.getProperty("execution_env").equalsIgnoreCase("local")) {
+			switch(browser.toLowerCase()) {
+			case "chrome" : driver = new ChromeDriver(); break;
+			case "edge" : driver = new EdgeDriver(); break;
+			case "firefox" : driver = new FirefoxDriver(); break;
+			default : System.out.println("Invalid browser"); return;
+			}
+		}
+		
+
 		
 		driver.manage().deleteAllCookies();
 		driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
